refactor(discussion): move discussion code from republik into it's own package, new util package
