+++ .env.example
@@ -148,6 +148,14 @@ AWS_SECRET_ACCESS_KEY=
 #PURGE_PSK=
 
 #############
+# discussions
+#############
+
+# the secret to hash userIds in discussions
+DISPLAY_AUTHOR_SECRET=RANDOM
+
+
+#############
 # github
 #############
