+++ packages/discussions/graphql/resolvers/Discussion/displayAuthor.js
