+++ packages/discussions/index.js
@@ -1,3 +1,4 @@
 module.exports = {
+  graphql: require('./graphql'),
   ...require('./lib')
 }
