+++ packages/discussions/lib/hotness.js
