+++ packages/discussions/loaders/Discussion.js
@@ -1,6 +1,15 @@
 const createDataLoader = require('@orbiting/backend-modules-dataloader')
 
 module.exports = (context) => ({
+  clear: async (id) => {
+    const discussion =
+      await context.loaders.Discussion.byId(id) ||
+      await context.loaders.Discussion.byRepoId(id)
+    if (discussion) {
+      context.loaders.Discussion.byId.clear(discussion.id)
+      context.loaders.Discussion.byRepoId.clear(discussion.repoId)
+    }
+  },
   byId: createDataLoader(ids =>
     context.pgdb.public.discussions.find({
       id: ids,
