+++ packages/utils/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@orbiting/backend-modules-discussions",
+  "name": "@orbiting/backend-modules-utils",
   "version": "0.0.1",
   "description": "let's discuss",
   "main": "index.js",
@@ -18,10 +18,5 @@
   "homepage": "https://github.com/orbiting/backends#readme",
   "scripts": {
     "link": "yarn link"
-  },
-  "dependencies": {
-    "@project-r/styleguide": "^5.81.7",
-    "dataloader": "^1.4.0",
-    "export-files": "^2.1.1"
   }
 }
