+++ servers/publikator/graphql/schema-types.js
@@ -211,4 +211,9 @@ enum OrderDirection {
   ASC
   DESC
 }
+
+# Credential stub for Discussion
+type Discussion {
+  dummy: Boolean
+}
 `
