+++ servers/republik/.env.example
@@ -10,9 +10,6 @@ CORS_WHITELIST_URL=http://localhost:3010,http://localhost:3003
 PUBLIC_WS_URL_BASE=ws://localhost:5000
 PUBLIC_WS_URL_PATH=/graphql
 
-# the secret to hash userIds in discussions
-DISPLAY_AUTHOR_SECRET=RANDOM
-
 # formet assets url, still used to detect old portrait urls
 ASSETS_BASE_URL=https://assets.republik.ch
